refac: folder archtecture of api module, separating routers per methods, aggregating services and models to your route
