+++ src/main/scala/api/controllers/bucket/Model.scala
@@ -7,12 +7,12 @@ import spray.json.{DefaultJsonProtocol, RootJsonFormat}
 
 
 final case class Bucket(
-                         name: String
+                  name: String
                        )
 
 final case class Buckets(
                   buckets:Seq[Bucket]
-                  )
+                        )
 
 
 trait BucketJsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
