+++ src/main/scala/api/controllers/bucket/exceptions/Exceptions.scala
@@ -1,5 +1,5 @@
 package pedro.goncalves
-package api.controllers.bucket
+package api.controllers.bucket.exceptions
 
 case class BucketNotExists(name:String) extends Exception
 case class BucketExists(name:String) extends Exception
