+++ src/main/scala/api/controllers/file/Model.scala
@@ -1,7 +1,7 @@
 package pedro.goncalves
 package api.controllers.file
 
-case class FileStorage(
+case class Model(
                       bucket:String,
                       repository:String,
                       versioned:Boolean,
